romio: fix misc conversion warnings
We promote the variable to larger integer type such as ADIO_Offset where
it is obvious. But some places are more entangled and one change will
lead to extensive changes. For example, the aggregation code. We use
(int) cast for now by adding necessary assertions. Future work should
properly extend these into large count interfaces.

+++ src/mpi/romio/adio/common/ad_coll_exch_new.c
@@ -39,8 +39,7 @@ void ADIOI_Print_flatlist_node(ADIOI_Flatlist_node * flatlist_node_p)
         fprintf(stderr, "print flatlist node of NULL ptr\n");
         return;
     }
-    fprintf(stderr, "print flatlist node count = %d (idx,blocklen)\n",
-            (int) flatlist_node_p->count);
+    fprintf(stderr, "print flatlist node count = %d (idx,blocklen)\n", flatlist_node_p->count);
     for (i = 0; i < flatlist_node_p->count; i++) {
         if (i % 5 == 0 && i != 0) {
             fprintf(stderr, "%d=(%lld,%lld)\n", i, (long long) flatlist_node_p->indices[i],
